Add newsletter signup
We are testing out MailChimp (free tier) for collecting, managing, and distributing newsletters to subscribers. Originally, I was going to create a custom sign-up form, but currently I need to do more research on the topic (XSS, accessibility and JS/React functionality). The first implementation is going to be an HTML file with an embedded MailChimp form.